common-ts: Fix really bad equality checks in eventual
_.isEqual returns true when comparing these two maps:

const m1 = new Map([['y','x']]);
const m2 = new Map([['x','y']]);

It also does that for random other maps, causing eventual values
that are _different_ not to propagate because it thought they are
equal. util.isDeepSstrictEqual doesn't have this problem.

This is a lodash issue from 2017 that is still unresolved:
https://github.com/lodash/lodash/issues/3428